Crop center the image in account info
Because we need to correctly crop and center the image we can no longer support snapping of toolbar in appbar. Since AppBarLayout is a subclass of LinearLayout and doesn't look deeper into subview hierachy for snapping children, thus it is better not to support it.

Also instead of using ScrollView and pad it with some weird magic numbers, use NestedScrollView class which is made to cooperate with AppBarLayout.